chore: move types to their own files
