+++ apps/api/src/app/metrics/services/average-metrics.types.ts
@@ -1,0 +1,6 @@
+export interface AverageMetricFilters {
+  workspaceId: number;
+  startDate: string;
+  endDate: string;
+  teamId: number;
+}
