+++ apps/api/src/app/metrics/services/chart-pull-request.types.ts
@@ -1,0 +1,9 @@
+import { Period } from "@sweetr/graphql-types/dist/api";
+
+export interface PullRequestChartFilters {
+  workspaceId: number;
+  startDate: string;
+  endDate: string;
+  period: Period;
+  teamId: number;
+}
